+++ src/tewi/service/search/yts_provider.py
@@ -297,6 +297,7 @@ class YTSProvider(BaseSearchProvider):
                 upload_date=upload_date,
                 provider=self.name,
                 provider_id=self.id(),
+                downloads=None,
                 page_url=page_url,
                 torrent_link=None,
                 fields=fields