+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add Prowlarr torrent search provider [#133](https://github.com/anlar/tewi/issues/133)
+- Display download count in search results [#137](https://github.com/anlar/tewi/issues/137)
 
 ### Changed